{"schema":"decodex/commit/1","summary":"Split MCP context","authority":"XY-1115"}
